requirements.txt
google-genai>=0.1.2
gradio>=4.20.0
pypdf>=4.1.0
python-docx>=1.1.0
fpdf2>=2.7.7
pydantic>=2.6.0
python-dotenv>=1.0.1
